GD Demonstration w/ Line
Created a line that slowly moves towards a user created dataset. This was very hard, it took a lot of time trying to come up with the right functions and implementation. Endless hours typing random numbers trying to find good learning rates to use, finally created a formula that used the user's input and scaled learning rates based on that. How it's possible to do it without user input data I have no idea but that's a problem for another day.